♻️ Del beregningsgrunnlag funksjoner i felles, V1 og V1 filer
