+++ src/main/kotlin/no/nav/tilleggsstonader/sak/vedtak/barnetilsyn/beregningV1/BeregningsgrunnlagUtils.kt
@@ -1,36 +1,91 @@
 package no.nav.tilleggsstonader.sak.vedtak.barnetilsyn.beregningV1
 
+import no.nav.tilleggsstonader.sak.felles.domain.BarnId
+import no.nav.tilleggsstonader.sak.vedtak.barnetilsyn.beregningFelles.BeregningsgrunnlagUtilsFelles
+import no.nav.tilleggsstonader.sak.vedtak.barnetilsyn.beregningV1.TilsynBeregningUtil.tilAktiviteterPerMånedPerType
+import no.nav.tilleggsstonader.sak.vedtak.barnetilsyn.beregningV1.TilsynBeregningUtil.tilDagerPerUke
+import no.nav.tilleggsstonader.sak.vedtak.barnetilsyn.beregningV1.TilsynBeregningUtil.tilUke
+import no.nav.tilleggsstonader.sak.vedtak.barnetilsyn.beregningV1.TilsynBeregningUtil.tilÅrMåned
+import no.nav.tilleggsstonader.sak.vedtak.barnetilsyn.domain.Aktivitet
+import no.nav.tilleggsstonader.sak.vedtak.barnetilsyn.domain.Beregningsgrunnlag
+import no.nav.tilleggsstonader.sak.vedtak.barnetilsyn.domain.StønadsperiodeGrunnlag
+import no.nav.tilleggsstonader.sak.vedtak.domain.StønadsperiodeBeregningsgrunnlag
+import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.domain.AktivitetType
+import kotlin.collections.component1
+import kotlin.collections.component2
+import kotlin.collections.map
+import kotlin.collections.sumOf
+
 object BeregningsgrunnlagUtils {
-    /**
-     * Beregner antall dager per uke som kan brukes
-     * Hvis antall dager fra stønadsperiode er 1, så kan man maks bruke 1 dag fra aktiviteter
-     * Hvis antall dager fra stønadsperiode er 5, men aktiviteter kun har 2 dager så kan man kun bruke 2 dager
-     */
-    fun beregnAntallAktivitetsdagerForUke(
-        periodeMedDager: PeriodeMedDager,
-        aktiviteter: List<PeriodeMedDager>,
-    ): Int =
-        aktiviteter.filter { it.overlapper(periodeMedDager) }.fold(0) { acc, aktivitet ->
-            // Tilgjengelige dager i uke i overlapp mellom stønadsperiode og aktivitet
-            val antallTilgjengeligeDager = minOf(periodeMedDager.antallDager, aktivitet.antallDager)
+    fun lagBeregningsgrunnlagPerMåned(
+        stønadsperioder: List<StønadsperiodeBeregningsgrunnlag>,
+        aktiviteter: List<Aktivitet>,
+        utgifterPerBarn: Map<BarnId, List<UtgiftBeregning>>,
+    ): List<Beregningsgrunnlag> {
+        val stønadsperioderPerMåned = stønadsperioder.tilÅrMåned()
+        val utgifterPerMåned = utgifterPerBarn.tilÅrMåned()
+        val aktiviteterPerMånedPerType = aktiviteter.tilAktiviteterPerMånedPerType()
 
-            trekkFraBrukteDager(periodeMedDager, aktivitet, antallTilgjengeligeDager)
-
-            acc + antallTilgjengeligeDager
+        return stønadsperioderPerMåned.entries.mapNotNull { (måned, stønadsperioder) ->
+            val aktiviteterForMåned = aktiviteterPerMånedPerType[måned] ?: error("Ingen aktiviteter for måned $måned")
+            utgifterPerMåned[måned]?.let { utgifter ->
+                val antallBarn = utgifter.map { it.barnId }.toSet().size
+                val makssats = finnMakssats(måned, antallBarn)
+                val stønadsperioderGrunnlag = finnStønadsperioderMedAktiviteter(stønadsperioder, aktiviteterForMåned)
+                Beregningsgrunnlag(
+                    måned = måned,
+                    makssats = makssats,
+                    stønadsperioderGrunnlag = stønadsperioderGrunnlag,
+                    utgifter = utgifter,
+                    utgifterTotal = utgifter.sumOf { it.utgift },
+                    antallBarn = antallBarn,
+                )
+            }
         }
+    }
 
-    /**
-     * Skal ikke kunne bruke en dager fra aktivitet eller stønadsperiode flere ganger.
-     * Trekker fra tilgjengelige dager fra antallDager
-     * Dersom stønadsperioden har 2 dager, og aktiviten 3 dager, så skal man kun totalt kunne bruke 2 dager
-     * Dersom stønadsperioden har 3 dager, og aktiviten 2 dager, så skal man kun totalt kunne bruke 2 dager
-     */
-    private fun trekkFraBrukteDager(
-        stønadsperiode: PeriodeMedDager,
-        aktivitet: PeriodeMedDager,
-        antallTilgjengeligeDager: Int,
-    ) {
-        aktivitet.antallDager -= antallTilgjengeligeDager
-        stønadsperiode.antallDager -= antallTilgjengeligeDager
+    private fun finnStønadsperioderMedAktiviteter(
+        stønadsperioder: List<StønadsperiodeBeregningsgrunnlag>,
+        aktiviteter: Map<AktivitetType, List<Aktivitet>>,
+    ): List<StønadsperiodeGrunnlag> {
+        val aktiviteterPerUke = aktiviteter.map { it.key to it.value.tilDagerPerUke() }.toMap()
+
+        return stønadsperioder.map { stønadsperiode ->
+            val relevanteAktiviteter = finnAktiviteterForStønadsperiode(stønadsperiode, aktiviteter)
+
+            StønadsperiodeGrunnlag(
+                stønadsperiode = stønadsperiode,
+                aktiviteter = relevanteAktiviteter,
+                antallDager = antallDager(stønadsperiode, aktiviteterPerUke),
+            )
+        }
+    }
+
+    private fun finnAktiviteterForStønadsperiode(
+        stønadsperiode: StønadsperiodeBeregningsgrunnlag,
+        aktiviteter: Map<AktivitetType, List<Aktivitet>>,
+    ): List<Aktivitet> =
+        aktiviteter[stønadsperiode.aktivitet]?.filter { it.overlapper(stønadsperiode) }
+            ?: error(
+                "Finnes ingen aktiviteter av type ${stønadsperiode.aktivitet} som passer med stønadsperiode med fom=${stønadsperiode.fom} og tom=${stønadsperiode.tom}",
+            )
+
+    private fun antallDager(
+        stønadsperiode: StønadsperiodeBeregningsgrunnlag,
+        aktiviteterPerType: Map<AktivitetType, Map<Uke, List<PeriodeMedDager>>>,
+    ): Int {
+        val stønadsperioderUker = stønadsperiode.tilUke()
+        val aktiviteterPerUke =
+            aktiviteterPerType[stønadsperiode.aktivitet]
+                ?: error("Finner ikke aktiviteter for ${stønadsperiode.aktivitet}")
+
+        return stønadsperioderUker
+            .map { (uke, periode) ->
+                val aktiviteterForUke =
+                    aktiviteterPerUke[uke]
+                        ?: error("Ingen aktivitet i uke fom=${uke.fom} og tom=${uke.tom}")
+
+                BeregningsgrunnlagUtilsFelles.beregnAntallAktivitetsdagerForUke(periode, aktiviteterForUke)
+            }.sum()
     }
 }
